Add splash screen, minor UI impovements

+++ lib/repositories/items_repository.dart
@@ -1,10 +1,18 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:firebase_auth/firebase_auth.dart';
 import 'package:simple_to_do_app/models/text_card_model.dart';
 
 class ItemsRepository {
   Stream<List<CardModel>> getItemStream() {
+    final userID = FirebaseAuth.instance.currentUser?.uid;
+    if (userID == null) {
+      throw Exception('User is not loged in');
+    }
     return FirebaseFirestore.instance
+        .collection('users')
+        .doc(userID)
         .collection('items')
+        .orderBy('checked')
         .snapshots()
         .map((querySnapshot) {
       return querySnapshot.docs.map((doc) {
@@ -15,14 +23,28 @@ class ItemsRepository {
 
   void changeCheckBoxValue(bool newcheckboxValue, String documentID) {
     final data = {'checked': newcheckboxValue};
+    final userID = FirebaseAuth.instance.currentUser?.uid;
+    if (userID == null) {
+      throw Exception('User is not loged in');
+    }
     FirebaseFirestore.instance
+        .collection('users')
+        .doc(userID)
         .collection('items')
         .doc(documentID)
         .set(data, SetOptions(merge: true));
   }
 
   Future<void> addNewTask(String task) async {
-    await FirebaseFirestore.instance.collection('items').add({
+    final userID = FirebaseAuth.instance.currentUser?.uid;
+    if (userID == null) {
+      throw Exception('User is not loged in');
+    }
+    await FirebaseFirestore.instance
+        .collection('users')
+        .doc(userID)
+        .collection('items')
+        .add({
       'checked': false,
       'title': task,
       'type': 'card',
@@ -30,7 +52,13 @@ class ItemsRepository {
   }
 
   Future<void> deleteTask({String? documentID}) {
+    final userID = FirebaseAuth.instance.currentUser?.uid;
+    if (userID == null) {
+      throw Exception('User is not loged in');
+    }
     return FirebaseFirestore.instance
+        .collection('users')
+        .doc(userID)
         .collection('items')
         .doc(documentID)
         .delete();
@@ -38,7 +66,13 @@ class ItemsRepository {
 
   void changeTaskText(String newTaskText, String documentID) {
     final data = {'title': newTaskText};
+    final userID = FirebaseAuth.instance.currentUser?.uid;
+    if (userID == null) {
+      throw Exception('User is not loged in');
+    }
     FirebaseFirestore.instance
+        .collection('users')
+        .doc(userID)
         .collection('items')
         .doc(documentID)
         .set(data, SetOptions(merge: true));
